+++ docs/app/components/Example/Simple/Simple.js
@@ -17,6 +17,7 @@ export default class Simple extends Component {
   state = { items: ITEMS }
 
   handleChange = (items) => {
+    console.log('change')
     this.setState({ items });
   }
 
@@ -27,7 +28,7 @@ export default class Simple extends Component {
         <h1 className="page-title">Simple</h1>
         <div className="row">
           <div className="col-12 col-lg-8 col-xl-6">
-            <Sortly items={items} itemRenderer={ItemRenderer} onChange={this.handleChange} />
+            <Sortly items={items} itemRenderer={ItemRenderer} onChange={this.handleChange} cancelOnDropOutside cancelOnDragOutside />
           </div>
         </div>
         <hr />